chore(rails 5): selectively reverting encoding
premailer-rails made changes after 1.9.4 that broke Planning Center's
usage within pco-communication. Encoding would flip-flop between
ASCII-8BIT and UTF-8, and when this was attempting to go over the wire
it could not be decoded by Planning Center People.

This commit acts as a minimal patch, that removes the parts we don't
need so emails continue working as expected in Planning Center
applications using pco-communication on Rails 5.

There is an integration test for this inside Planning Center Giving. I
tried to write a test that could be generic and live in this fork of the
gem, but ultimately the entire system had to be exercised by hitting
`people.pco.dev` to encounter the error.

My hope is that by consolidating this patch to one commit, we can stay
up-to-date with premailer-rails as the project evolves, with minimal
maintenance.

Original work to get this problem solved was by @zhubert. Thanks Zack!

Moving to ministrycentered/premailer-rails and basing off of 1.9.7 by
@danott